Refactor archived component into its own file
