+++ src/Practitioner/PatientProfile/index.js
@@ -12,10 +12,9 @@ import TreatmentTimeline from '../../Basics/TreatmentTimeline'
 import ReportingHistory from './ReportingHistory'
 import SectionLabel from '../../Components/SectionLabel'
 import PatientProfileDialogs from './Dialogs'
-import ErrorOutlineIcon from '@material-ui/icons/ErrorOutline';
-import { Grid, Typography } from '@material-ui/core';
 import HorizontalButtons from './HorizontalOptions'
 import Avatar from '@material-ui/core/Avatar'
+import ArchivedOptions from './ArchivedOptions';
 
 //Styles are at the bottom :)
 
@@ -50,7 +49,6 @@ const Profile = observer((props) => {
             <PatientProfileDialogs />
             {patientProfileStore.selectedPatient.loaded ?
                 <>{!patientProfileStore.selectedPatient.accessError ? <div className={classes.patientContainer}>
-                    {patientProfileStore.isArchived && <ArchivedError />}
                     <div className={classes.header}>
                         <div className={classes.profileHeader}>
                             <Avatar style={{ backgroundColor: Colors.green, marginRight: "1em" }} size="small">{patientProfileStore.selectedPatient.details.fullName[0]}</Avatar>
@@ -58,7 +56,7 @@ const Profile = observer((props) => {
                         </div>
                         <HorizontalButtons />
                     </div>
-                    
+                    <ArchivedOptions />
                     <div className={classes.top}>
                         <PatientInfo />
                         <TreatmentStatus />
@@ -75,19 +73,6 @@ const Profile = observer((props) => {
 
         </>)
 });
-
-
-const ArchivedError = () => {
-    const { t } = useTranslation('translation');
-    const classes = useStyles();
-
-    return (
-        <Grid container alignItems="center" justify="center" className={classes.archived}>
-            <ErrorOutlineIcon />
-            <Typography variant="body1">{t('archive.warningShort')}</Typography>
-        </Grid>)
-}
-
 
 const Loading = () => {
     const classes = useStyles();
@@ -140,18 +125,6 @@ const useStyles = makeStyles({
         width: "100%",
         height: "100%",
         ...Styles.flexCenter
-    },
-    archived: {
-        width: "100%",
-        backgroundColor: Colors.warningRed,
-        color: "white",
-        padding: "1em",
-        "& > p": {
-            marginLeft: ".5em"
-        },
-        borderTopLeftRadius: "4px",
-        borderTopRightRadius: "4px"
-
     },
     header: {
         width: "100%",
